bug: Improve handling exceptions from unsupported PHPUnit versions
Attempting to render PHPUnit assertion failures is brittle, because
a) we (intentionally) don't impose any PHPUnit version constraints
in composer.json and b) PHPUnit's code for rendering exceptions is
not covered by their BC promise.

This update to the PHPUnitExceptionStringer ensures that we always
render something useful - prioritising communicating that an
assertion has failed - even if we don't correctly support the
PHPUnit version in use.