refactor: Tighten types for ContextFactory
Since it was introduced in e16301ef in 2014, ContextFactory has had a
strict `Context $context` typehint on the argument to
`initializeInstance`.

This means that although theoretically it could have taken any `$class`
name to `createContext`, and returned any object from `createInstance`,
in practice this would have failed on the type of argument passed to
`initializeInstance`.

Therefore, it is not a BC break to be strict that the original `$class`
arg passed into `createContext` must be a `class-string<Context>` and
the return of `createInstance` must be `Context`.