refactor: Improve return type info for `InteractiveContextIdentifier`
Rector gave `askQuestion` a mixed return type, because the symfony
method has a mixed return type. However, the actual return type depends
entirely on the question you pass in (for example, if you register
validators or normalizers to cast the answer, or provide choices that
are not strings).

If you pass in a list of string choices (as we do) then you will always
receive a string answer. We can therefore be strict about this return
type, which allows us to also be strict about the return of the public
method.